Rename Checkout to CreditCardCheckout
